refactor: Change hashtag dto path
